Changed some things in the help panel Added guildinfo and botinfo Added a limit to purge

+++ cogs/server.py
@@ -331,8 +331,12 @@ class Server(commands.Cog):
     @commands.cooldown(rate = 1, per = 5.0)
     @commands.has_permissions(manage_messages = True)
     async def purge(self, ctx, amount = 5):
-        await ctx.message.delete()
-        await ctx.channel.purge(limit = amount)
+        if amount < 25:
+
+            await ctx.message.delete()
+            await ctx.channel.purge(limit = amount)
+        else:
+            await ctx.send('The maximum amount of messages you can purge is 25!')
